refactor: store typings and observation helpers in separate files
